+++ index.html
@@ -1,11 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Card Memory</title>
-  <link rel="stylesheet" href="style.css" />
+  <link rel="stylesheet" href="style/style.css" />
 </head>
+
 <body>
   <main>
     <header>
@@ -40,6 +42,7 @@
     </div>
   </main>
 
-  <script src="script.js"></script>
+  <script type="module" src="scripts/main.js"></script>
 </body>
+
 </html>
