+++ style/style.css
@@ -1,13 +1,15 @@
 @font-face {
   font-family: "Gilroy";
-  src: url(./67494711d21558f07d259e65_Gilroy-Regular.ttf);
+  src: url(../assets/fonts/67494711d21558f07d259e65_Gilroy-Regular.ttf);
 }
+
 @font-face {
   font-family: "Dharam";
-  src: url(./DharmaGothicE-Bold.ttf);
+  src: url(../assets/fonts/DharmaGothicE-Bold.ttf);
 }
+
 * {
-  margin: 0;  
+  margin: 0;
   padding: 0;
   box-sizing: border-box;
   font-family: Gilroy, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
@@ -21,11 +23,13 @@ body {
   justify-content: center;
   align-items: center;
 }
+
 html.light-mode,
 body.light-mode {
   background: linear-gradient(135deg, #667eea, #764ba2);
   color: black;
 }
+
 html.dark-mode,
 body.dark-mode {
   background: #121212;
@@ -51,25 +55,30 @@ main {
   text-align: center;
   color: #fff;
 }
+
 main h1 {
   font-size: 6rem;
   margin-bottom: 2rem;
   letter-spacing: 1px;
 }
+
 main .screen {
   width: 100%;
   height: 100%;
 }
+
 main .screen .part-1,
 main .screen .part-3 {
   width: 100%;
   height: 100%;
 }
+
 main .screen .part-1 h2,
 main .screen .part-3 h2 {
   font-family: Dharam;
   font-size: 4rem;
 }
+
 main .screen .part-2 .card-list {
   display: grid;
   grid-template-columns: repeat(6, 100px);
@@ -78,10 +87,12 @@ main .screen .part-2 .card-list {
   list-style: none;
   padding: 0;
 }
+
 main .screen .part-2 .card-list .card {
   cursor: pointer;
   perspective: 1000px;
 }
+
 main .screen .part-2 .card-list .card .card-inner {
   position: relative;
   width: 100px;
@@ -89,9 +100,11 @@ main .screen .part-2 .card-list .card .card-inner {
   transform-style: preserve-3d;
   transition: transform 0.6s;
 }
+
 main .screen .part-2 .card-list .card.flipped .card-inner {
   transform: rotateY(180deg);
 }
+
 main .screen .part-2 .card-list .card .card-front,
 main .screen .part-2 .card-list .card .card-back {
   position: absolute;
@@ -106,17 +119,21 @@ main .screen .part-2 .card-list .card .card-back {
   backface-visibility: hidden;
   box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2);
 }
+
 main .screen .part-2 .card-list .card .card-front {
   background: #e2e8f0;
 }
+
 main .screen .part-2 .card-list .card .card-back {
   background: #38b2ac;
   color: #fff;
   transform: rotateY(180deg);
 }
+
 main .screen .part-3 h2 {
   font-size: 12rem;
 }
+
 main .win-screen {
   z-index: -1;
   position: fixed;
@@ -130,10 +147,12 @@ main .win-screen {
   justify-content: center;
   align-items: center;
 }
+
 main .win-screen div h1 {
   font-size: 6rem;
   text-transform: uppercase;
 }
+
 main .win-screen div button {
   padding: 1rem 2rem;
   border-radius: 5px;
@@ -144,9 +163,11 @@ main .win-screen div button {
   font-weight: bolder;
   font-size: x-large;
 }
+
 main .win-screen div button:hover {
   background: #287e79;
 }
+
 main .win-screen div button:active {
   scale: 0.95;
 }
@@ -159,25 +180,30 @@ main .win-screen div button:active {
     text-align: center;
     color: #fff;
   }
+
   main h1 {
     font-size: 3rem;
     margin-top: 6rem;
     letter-spacing: 0.5px;
   }
+
   main .screen {
     width: 100%;
     height: 100%;
   }
+
   main .screen .part-1,
   main .screen .part-3 {
     width: 100%;
     height: 100%;
   }
+
   main .screen .part-1 h2,
   main .screen .part-3 h2 {
     font-family: Dharam;
     font-size: 2rem;
   }
+
   main .screen .part-2 .card-list {
     display: grid;
     grid-template-columns: repeat(3, 100px);
@@ -186,19 +212,23 @@ main .win-screen div button:active {
     list-style: none;
     padding: 0 1rem;
   }
+
   main .screen .part-2 .card-list .card {
     cursor: pointer;
     perspective: 1000px;
   }
+
   main .screen .part-2 .card-list .card .card-inner {
     position: relative;
     width: 80px;
     height: 80px;
     transition: transform 0.6s;
   }
+
   main .screen .part-2 .card-list .card.flipped .card-inner {
     transform: rotateY(180deg);
   }
+
   main .screen .part-2 .card-list .card .card-front,
   main .screen .part-2 .card-list .card .card-back {
     position: absolute;
@@ -213,17 +243,21 @@ main .win-screen div button:active {
     backface-visibility: hidden;
     box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2);
   }
+
   main .screen .part-2 .card-list .card .card-front {
     background: #e2e8f0;
   }
+
   main .screen .part-2 .card-list .card .card-back {
     background: #38b2ac;
     color: #fff;
     transform: rotateY(180deg);
   }
+
   main .screen .part-3 h2 {
     font-size: 4rem;
   }
+
   main .win-screen {
     z-index: -1;
     position: fixed;
@@ -238,10 +272,12 @@ main .win-screen div button:active {
     align-items: center;
     padding: 1rem;
   }
+
   main .win-screen div h1 {
     font-size: 2.5rem;
     text-transform: uppercase;
   }
+
   main .win-screen div button {
     padding: 1rem 2rem;
     border-radius: 5px;
@@ -253,10 +289,14 @@ main .win-screen div button:active {
     font-size: 1rem;
     padding: 0.75rem 1.5rem;
   }
+
   main .win-screen div button:hover {
     background: #287e79;
   }
+
   main .win-screen div button:active {
     scale: 0.95;
   }
-}/*# sourceMappingURL=style.css.map */
+}
+
+/*# sourceMappingURL=style.css.map */
